drm/msm/dsi-staging: Disable partial display updates on all platforms
Due to the misleading nature of the original commit message, I was
operating under the false pretense that enabling this feature would
benefit my device (OnePlus 7T). Turns out, that's not the case.
Coupled with the fact that this historically causes issues with
particular apps and creates jank when used in conjunction with DC
dimming on platforms which do support the feature, I see no reason
to continue enabling partial update support for anyone.

This reverts the following commits:
daa7886 ("drm/msm/dsi-staging: enable partial updates only for S6E3HC2")
6925ed9 ("drm/msm/dsi-staging: enable partial updates")

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>