Revert "neutrino: Reduce (DEVFREQ_)INPUT_BOOST_DURATION_MS"
Our in-kernel boosting framework relies on sched_boost for duration
timing, custom ROMs generally do not use CAF's userspace HAL, and
thus do not support sched_boost. Because of this, we need to re-set
a base duration so that the kernel at least behaves in a close
aproximation to how it would on OOS. Otherwise, we may as well not
have input boosting at all.

This reverts commit 00c73f97da6cc6c41ea6766a64f28779bdb2ff8f.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>